add complete profile modal to all create project and user dashboard routes
