+++ src/components/Header/Header.sc.tsx
@@ -1,22 +1,18 @@
 import styled from 'styled-components';
-import { Row } from '@/components/styled-components/Grid';
-import { Button as CButton } from '@/components/styled-components/Button';
-import { IHeader } from './Header';
 import {
 	neutralColors,
 	brandColors,
 	Button,
 	GLink,
-	Subline,
 	ButtonLink,
 } from '@giveth/ui-design-system';
+
+import { Row } from '@/components/styled-components/Grid';
+import { Button as CButton } from '@/components/styled-components/Button';
+import { IHeader } from './Header';
 import { device } from '@/utils/constants';
 import { ETheme } from '@/context/general.context';
 import { Shadow } from '@/components/styled-components/Shadow';
-
-export const HeaderPlaceholder = styled.div`
-	height: 100px;
-`;
 
 export const StyledHeader = styled(Row)<IHeader>`
 	position: fixed;
@@ -87,13 +83,6 @@ export const HBContainer = styled.div`
 	z-index: 2;
 `;
 
-export const MBContainer = styled(HBContainer)`
-	display: flex;
-	align-items: center;
-	z-index: 2;
-	gap: 8px;
-`;
-
 export const WBInfo = styled.div`
 	display: flex;
 	flex-direction: column;
@@ -155,22 +144,6 @@ export const HeaderLinks = styled(Row)<IThemed>`
 	}
 `;
 
-export const SmallHeaderLinks = styled(Row)`
-	align-self: center;
-	align-items: center;
-	display: flex;
-	flex-direction: column;
-	padding: 12px 16px 12px 48px;
-	border-radius: 24px;
-	gap: 8px;
-	position: relative;
-	background-color: ${props =>
-		props.theme === ETheme.Dark ? brandColors.giv[900] : 'white'};
-	border: 1px solid
-		${props =>
-			props.theme === ETheme.Dark ? brandColors.giv[600] : 'white'};
-`;
-
 export const HeaderLink = styled(GLink)<IHeaderLinkProps>`
 	padding: 8px 16px 7px;
 	border-radius: 72px;
@@ -195,24 +168,6 @@ export const ConnectButton = styled(Button)`
 		props.theme === ETheme.Dark ? '' : Shadow.Dark[500]};
 `;
 
-export const NotifButton = styled(HeaderButton)`
-	padding: 23px;
-	background-image: url('/images/notif.svg');
-	background-position: center;
-	background-repeat: no-repeat;
-	max-width: 48px;
-`;
-
-export const CreateProject = styled(ButtonLink)`
-	white-space: nowrap;
-	display: none;
-	box-shadow: ${props =>
-		props.theme === ETheme.Dark ? '' : Shadow.Dark[500]};
-	@media ${device.laptop} {
-		display: flex;
-	}
-`;
-
 export const SmallCreateProject = styled(ButtonLink)`
 	white-space: nowrap;
 	padding: 0;
@@ -232,12 +187,6 @@ export const SmallCreateProjectParent = styled.div`
 	}
 `;
 
-export const BalanceTooltip = styled(Subline)`
-	color: ${neutralColors.gray[100]};
-	text-align: center;
-	width: 120px;
-`;
-
 export const MenuAndButtonContainer = styled.div`
 	position: relative;
 	z-index: 2;
@@ -253,18 +202,4 @@ export const CoverLine = styled.div<IThemed>`
 	top: 1px;
 	bottom: 4px;
 	border-radius: 48px;
-`;
-
-export const IconMenuWrapper = styled.div`
-	position: absolute;
-	left: 16px;
-	top: 12px;
-	color: ${brandColors.giv[500]};
-`;
-export const HeaderSmallMenuAndButtonContainer = styled(MenuAndButtonContainer)`
-	position: relative;
-	z-index: 2;
-	@media ${device.laptopL} {
-		display: none;
-	}
 `;
