+++ src/components/views/projects/ProjectsIndex.tsx
@@ -16,7 +16,11 @@ import { BigArc } from '@/components/styled-components/Arc';
 import ProjectCard from '@/components/project-card/ProjectCard';
 import SearchBox from '@/components/SearchBox';
 import Routes from '@/lib/constants/Routes';
-import { capitalizeFirstLetter, mediaQueries } from '@/lib/helpers';
+import {
+	capitalizeFirstLetter,
+	isUserRegistered,
+	mediaQueries,
+} from '@/lib/helpers';
 import { FETCH_ALL_PROJECTS } from '@/apollo/gql/gqlProjects';
 import { initializeApollo } from '@/apollo/apolloClient';
 import { ICategory, IProject } from '@/apollo/types/types';
@@ -75,10 +79,12 @@ const buildCategoryObj = (array: ICategory[]) => {
 };
 
 const ProjectsIndex = (props: IProjectsView) => {
+	const { projects, totalCount: _totalCount, categories } = props;
+
 	const {
 		state: { user },
+		actions: { showCompleteProfile },
 	} = useUser();
-	const { projects, totalCount: _totalCount, categories } = props;
 
 	const [categoriesObj, setCategoriesObj] = useState<ISelectObj[]>();
 	const [selectedCategory, setSelectedCategory] =
@@ -176,6 +182,14 @@ const ProjectsIndex = (props: IProjectsView) => {
 		pageNum.current = pageNum.current + 1;
 	};
 
+	const handleCreateButton = () => {
+		if (isUserRegistered(user)) {
+			router.push(Routes.CreateProject);
+		} else {
+			showCompleteProfile();
+		}
+	};
+
 	const showLoadMore = totalCount > filteredProjects.length;
 
 	return (
@@ -245,7 +259,7 @@ const ProjectsIndex = (props: IProjectsView) => {
 							}
 						/>
 						<StyledButton
-							onClick={() => router.push(Routes.CreateProject)}
+							onClick={handleCreateButton}
 							label='Create a Project'
 							transparent
 						/>
